+++ src/Devices/CachedCollection.php
@@ -43,26 +43,12 @@ final class CachedCollection implements CollectionInterface
     }
 
 
-    /**
-     * Set the logger object to use.
-     *
-     * @var LoggerInterface $logger The logging object
-     *
-     * @return $this
-     */
-    public function setLogger(LoggerInterface $logger)
+    public function setLogger(LoggerInterface $logger): void
     {
         $this->collection->setLogger($logger);
-
-        return $this;
     }
 
 
-    /**
-     * Get the logger object to use.
-     *
-     * @return LoggerInterface $logger The logging object
-     */
     public function getLogger(): LoggerInterface
     {
         return $this->collection->getLogger();