+++ src/Network.php
@@ -60,27 +60,13 @@ final class Network implements NetworkInterface, LoggerAwareInterface
     }
 
 
-    /**
-     * Set the logger object to use.
-     *
-     * @var LoggerInterface $logger The logging object
-     *
-     * @return $this
-     */
-    public function setLogger(LoggerInterface $logger)
+    public function setLogger(LoggerInterface $logger): void
     {
         $this->collection->setLogger($logger);
-
-        return $this;
     }
 
 
-    /**
-     * Get the logger object to use.
-     *
-     * @return LoggerInterface $logger The logging object
-     */
-    public function getLogger()
+    public function getLogger(): LoggerInterface
     {
         return $this->collection->getLogger();
     }