+++ tests/NetworkTest.php
@@ -37,7 +37,8 @@ class NetworkTest extends MockTest
     {
         $logger = Mockery::mock(LoggerInterface::class);
         $this->collection->shouldReceive("setLogger")->with($logger)->once();
-        $this->assertSame($this->network, $this->network->setLogger($logger));
+        $this->network->setLogger($logger);
+        $this->assertTrue(true);
     }